Move project files into their own subdirectory
