+++ lib/mobilis.rb
@@ -16,13 +16,13 @@ end
 require_relative "mobilis/version"
 require_relative "mobilis/logger"
 
-require_relative "mobilis/generic_project"
 require_relative "mobilis/project"
-require_relative "mobilis/localgem_project"
-require_relative "mobilis/redis_instance"
-require_relative "mobilis/mysql_instance"
-require_relative "mobilis/postgresql_instance"
-require_relative "mobilis/rack_project"
+require_relative "mobilis/project/generic_project"
+require_relative "mobilis/project/localgem_project"
+require_relative "mobilis/project/redis_instance"
+require_relative "mobilis/project/mysql_instance"
+require_relative "mobilis/project/postgresql_instance"
+require_relative "mobilis/project/rack_project"
 
 require_relative "mobilis/actions_projects_take"
 require_relative "mobilis/command_line"
