+++ lib/mobilis/project/generic_project.rb
