+++ lib/mobilis/project/localgem_project.rb
