+++ lib/mobilis/project/rails_project.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "fileutils"
 
 require "mobilis/generic_project"
 require "mobilis/os"
@@ -193,6 +192,7 @@ module Mobilis
     end
 
     def wait_until_line
+      # TODO FIXME 
       if database.instance_of? Mobilis::PostgresqlInstance
         return <<~POSTGRES_LINE
           /myapp/wait-until "psql postgres://#{database.username}:#{database.password}@#{database.name}/#{name}_production -c 'select 1'"
